refractor file names and removed access token class
