+++ apps/funding-platform/.env
@@ -1,2 +1,4 @@
+# Secret key used for access token generation
 SECRET_KEY=
-PRIV_KEY=
+# Wallet private key that will be completing the requests
+WALLET_PRIV_KEY=
