+++ apps/funding-platform/package.json
@@ -16,7 +16,6 @@
     "@nomiclabs/hardhat-ethers": "^2.2.1",
     "@types/express": "^4.17.14",
     "@types/supertest": "^2.0.12",
-    "dotenv": "^16.0.3",
     "rpch-configs-eslint": "*",
     "rpch-configs-jest": "*",
     "rpch-configs-prettier": "*",
@@ -30,9 +29,6 @@
     "express": "^4.18.2",
     "hardhat": "^2.12.3",
     "lowdb": "^5.0.5",
-    "mysql2": "^2.3.3",
-    "rpch-common": "*",
-    "serverless-mysql": "^1.5.4",
-    "sql-template-strings": "^2.2.2"
+    "rpch-common": "*"
   }
 }
