+++ apps/funding-platform/src/utils/index.ts
@@ -1,0 +1,6 @@
+export const isExpired = (expireAt: string) => {
+  if ((new Date(expireAt).valueOf() ?? 0) < new Date(Date.now()).valueOf()) {
+    return true;
+  }
+  return false;
+};
